pyproject.toml
[tool.poetry]
name = "natureremo"
version = "0.1.0"
description = ""
authors = ["Katsunori Fukui <katsunori.fukui04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
nature-remo = "^0.3.3"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
